+++ src/plugin/Routes.php
@@ -14,10 +14,10 @@ trait Routes
     {
         Event::on(UrlManager::class, UrlManager::EVENT_REGISTER_CP_URL_RULES, function(RegisterUrlRulesEvent $event) {
             $event->rules['colorit'] = ['template' => 'colorit/index'];
-            $event->rules['colorit/settings/general'] = 'colorit/settings/general';
-            $event->rules['colorit/settings/presets'] = 'colorit/presets/index';
-            $event->rules['colorit/settings/presets/<presetId:\d+>'] = 'colorit/presets/edit';
-            $event->rules['colorit/settings/presets/new'] = 'colorit/presets/edit';
+            $event->rules['colorit/settings'] = 'colorit/settings';
+            $event->rules['colorit/presets'] = 'colorit/presets/index';
+            $event->rules['colorit/presets/<presetId:\d+>'] = 'colorit/presets/edit';
+            $event->rules['colorit/presets/new'] = 'colorit/presets/edit';
         });
     }
 }